Consolidate link and file preview components
changed: merge LinkPreview and FilePreview into a single
AttachmentPreview component
changed: abstract direct media handling and URL scraping logic
changed: integrate site mapping and URL filtering into the new component
changed: add support for local media previews, including images and PDFs
changed: improve handling of YouTube embeds and general website previews
fix: ensure correct rendering of various attachment types in messages
